05_hackathon/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
shiny>=0.4.0
pandas>=1.5.0
plotly>=5.14.0
shinywidgets>=0.3.0
